Update README.md
The Accounts web app serves as a comprehensive backend solution designed to highlight the intricacies of user authentication and API development within a Django framework. By leveraging the powerful features of Strawberry GraphQL, the application offers a seamless and efficient means of interacting with user data through GraphQL APIs.

The project's emphasis on the Base User Model underscores the importance of tailored user management functionalities, allowing for a more personalized and flexible approach to user authentication and authorization. Through the utilization of JWT authentication, the security and integrity of user data are ensured, enabling secure access to the application's features and resources.

Developed with a keen focus on demonstrating proficiency in GraphQL-based API development, the Accounts web app showcases best practices in structuring and querying data, making it an ideal platform for exploring the capabilities of GraphQL in a real-world scenario. By providing a clear and concise interface for interacting with user data, the application simplifies the process of managing user accounts and accessing relevant information efficiently.